feat(die): share tracking context
