+++ .changeset/shared-die-tracking-context.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/live-dmk-shared": patch
+---
+
+Export the shared Device Intent Executor tracking context.
