+++ libs/live-dmk-shared/README.md
@@ -14,6 +14,8 @@ Platform-agnostic shared logic for the Ledger Device Management Kit (DMK) integr
   - `LedgerLiveLogger` — logging adapter bridging DMK events to Ledger Live's log system
   - `LiveBlindSigningReporter` — reports blind-signing events for analytics/safety
   - `UserHashService` — hashed user identifier for DMK telemetry
+- `DeviceIntentTrackingProvider` / `useDeviceIntentTracking` — shared context contract for
+  platform-specific Device Intent Executor tracking
 - Centralises DMK configuration
 
 ## Key exports / concepts
