+++ libs/live-dmk-shared/src/index.ts
@@ -30,6 +30,15 @@ export {
 } from "./connectDevice/connectDeviceUseCase";
 export { activeDeviceSessionSubject } from "./config/activeDeviceSession";
 export { dmkToLedgerDeviceIdMap, ledgerToDmkDeviceIdMap } from "./config/dmkToLedgerDeviceIdMap";
+export {
+  DeviceIntentTrackingProvider,
+  useDeviceIntentTracking,
+} from "./deviceIntentTracking/DeviceIntentTrackingContext";
+export type {
+  DeviceIntentTrackingContextValue,
+  DeviceIntentTrackingProperties,
+  SourceFlow,
+} from "./deviceIntentTracking/DeviceIntentTrackingContext";
 export { LedgerLiveLogger } from "./services/LedgerLiveLogger";
 export { UserHashService } from "./services/UserHashService";
 export {
